Extract `AnyPacker` tests
Also:
  * Improve method name for ensuring unpacked message.
